requirements.txt
pandas==1.2.4
instagramy==4.4
requests==2.21.0
beautifulsoup4==4.9.3
Pillow==8.3.0
